fix: unpack and download were conflicting
done would get triggered on complete and not on unpack.
to prevent conflicts
