+++ src/electron/future/ipc/install.ts
@@ -7,55 +7,73 @@ import { appSettingsStore } from "@/stores";
 import { getCaptainData, getCaptainDownloads, getDirectory } from "@/utils/path-helpers";
 import { unpack } from "@/utils/unpack";
 
-ipcMain.on(buildKey([ID.INSTALL], { suffix: ":start" }), async (_event, data) => {
-	const window_ = BrowserWindow.getFocusedWindow();
-	if (!window_) {
-		return;
-	}
-
-	try {
-		// Iterate over each download object
-		for (const { url, destination } of data) {
-			await download(window_, url, {
-				directory: getCaptainDownloads(),
-
-				onStarted() {
-					appSettingsStore.set("status", DownloadState.ACTIVE);
-					window_.webContents.send(buildKey([ID.INSTALL], { suffix: ":started" }), true);
-				},
-				onProgress(progress) {
-					window_.webContents.send(
-						buildKey([ID.INSTALL], { suffix: ":progress" }),
-						progress
-					);
-				},
-				onCancel() {
-					window_.webContents.send(
-						buildKey([ID.INSTALL], { suffix: ":cancelled" }),
-						true
-					);
-					appSettingsStore.set("status", DownloadState.CANCELLED);
-				},
-				async onCompleted(item) {
-					window_.webContents.send(
-						buildKey([ID.INSTALL], { suffix: ":unpacking" }),
-						true
-					);
-					appSettingsStore.set("status", DownloadState.UNPACKING);
-
-					const targetPath = getCaptainData(destination);
-					await unpack(getDirectory("7zip", "win", "7za.exe"), item.path, targetPath);
-				},
-			});
+ipcMain.on(
+	buildKey([ID.INSTALL], { suffix: ":start" }),
+	async (_event, data: { url: string; destination: string; unzip?: boolean }[]) => {
+		const window_ = BrowserWindow.getFocusedWindow();
+		if (!window_) {
+			return;
 		}
 
-		// Everything was downloaded & unpacked
-		window_.webContents.send(buildKey([ID.INSTALL], { suffix: ":completed" }), true);
-		appSettingsStore.set("status", DownloadState.DONE);
-	} catch (error) {
-		if (error instanceof Error) {
-			window_.webContents.send(buildKey([ID.INSTALL], { suffix: ":failed" }), error.message);
-			appSettingsStore.set("status", DownloadState.FAILED);
+		try {
+			// Iterate over each download object
+			const items: Promise<void>[] = [];
+			for (const { url, destination, unzip } of data) {
+				await download(window_, url, {
+					directory: getCaptainDownloads(),
+
+					onStarted() {
+						appSettingsStore.set("status", DownloadState.ACTIVE);
+						window_.webContents.send(
+							buildKey([ID.INSTALL], { suffix: ":started" }),
+							true
+						);
+					},
+					onProgress(progress) {
+						window_.webContents.send(
+							buildKey([ID.INSTALL], { suffix: ":progress" }),
+							progress
+						);
+					},
+					onCancel() {
+						window_.webContents.send(
+							buildKey([ID.INSTALL], { suffix: ":cancelled" }),
+							true
+						);
+						appSettingsStore.set("status", DownloadState.CANCELLED);
+					},
+					async onCompleted(file) {
+						const targetPath = getCaptainData(destination);
+
+						if (unzip) {
+							// Unpack immediately and send to array to allow awaiting multiple unzips
+							items.push(
+								unpack(
+									getDirectory("7zip", "win", "7za.exe"),
+									file.path,
+									targetPath
+								)
+							);
+						}
+					},
+				});
+			}
+
+			// Set the state to unpacking and wait until all items are unpacked and ready
+			window_.webContents.send(buildKey([ID.INSTALL], { suffix: ":unpacking" }), true);
+			appSettingsStore.set("status", DownloadState.UNPACKING);
+			await Promise.all(items);
+			// Everything was downloaded & unpacked
+			window_.webContents.send(buildKey([ID.INSTALL], { suffix: ":completed" }), true);
+			appSettingsStore.set("status", DownloadState.DONE);
+		} catch (error) {
+			if (error instanceof Error) {
+				window_.webContents.send(
+					buildKey([ID.INSTALL], { suffix: ":failed" }),
+					error.message
+				);
+				appSettingsStore.set("status", DownloadState.FAILED);
+			}
 		}
 	}
-});
+);
